Time: 31 ms (73.47%), Space: 15.5 MB (89.89%) - LeetHub

+++ 3195-separate-black-and-white-balls/3195-separate-black-and-white-balls.cpp
@@ -5,7 +5,7 @@ public:
         int black=0;
         for(int i=0;i<s.length();i++){
             if(s[i]=='0'){
-                swap+=black;
+                swap+=(long long)black;
 
             }
             else{
